+++ SharpSource/SharpSource.Test/StringConcatenatedInLoopTests.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-
+using SharpSource.Test.Helpers;
 using VerifyCS = SharpSource.Test.CSharpCodeFixVerifier<SharpSource.Diagnostics.StringConcatenatedInLoopAnalyzer, Microsoft.CodeAnalysis.Testing.EmptyCodeFixProvider>;
 
 namespace SharpSource.Test;
@@ -151,6 +151,53 @@ while (true)
 var res = string.Empty;
 while (true)
     {|#0:res += ""test""|};
+";
+
+        await VerifyCS.VerifyDiagnosticWithoutFix(original, VerifyCS.Diagnostic().WithMessage("A string was concatenated in a loop which introduces intermediate allocations. Consider using a StringBuilder or pre-allocated string instead."));
+    }
+
+    [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/290")]
+    public async Task StringConcatenatedInLoopTests_NoConcatenation()
+    {
+        var original = @"
+var res = string.Empty;
+while (true)
+{
+    res = ""hello"";
+}
+";
+
+        await VerifyCS.VerifyNoDiagnostic(original);
+    }
+
+    [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/290")]
+    public async Task StringConcatenatedInLoopTests_PropertyAssignment()
+    {
+        var original = @"
+Test res = null;
+while (true)
+{
+    res = new Test { MyProp = ""hello"" };
+}
+
+class Test
+{
+    public string MyProp { get; set; }
+}
+";
+
+        await VerifyCS.VerifyNoDiagnostic(original);
+    }
+
+    [TestMethod]
+    public async Task StringConcatenatedInLoopTests_AssignmentAndConcatenationSeparated()
+    {
+        var original = @"
+var res = string.Empty;
+while (true)
+{
+    {|#0:res = res + ""test""|};
+}
 ";
 
         await VerifyCS.VerifyDiagnosticWithoutFix(original, VerifyCS.Diagnostic().WithMessage("A string was concatenated in a loop which introduces intermediate allocations. Consider using a StringBuilder or pre-allocated string instead."));