+++ CHANGELOG.md
@@ -3,6 +3,7 @@ https://keepachangelog.com/en/1.0.0/
 
 ## [1.21.1] - 2023-01-21
 - `StringPlaceholdersInWrongOrder` and `UnnecessaryEnumerableMaterialization`: Rewritten to use `IOperation`
+- `UnnecessaryEnumerableMaterialization` tests use `VerifyCS`
 
 ## [1.21.0] - 2023-01-21
 - `StringConcatenatedInLoop`: A `string` was concatenated in a loop which introduces intermediate allocations. Consider using a `StringBuilder` or pre-allocated `string` instead.